+++ src/components/Prayers/Prayers.js
@@ -8,7 +8,7 @@ import Prayer from "./Prayer";
 import "./prayers.css";
 import Timer from "../Timer";
 import { getJustPrayers } from "../../utils";
-import { useRenderCounts, useWhatsapplogger } from "../../hooks/api-hooks";
+import { useRenderCounts, useWhatsapplogger, useHijriHook } from "../../hooks/api-hooks";
 import Accordion from "../Accordion";
 import Metadata from "./Metadata";
 import TodayIcon from "@material-ui/icons/Today";
@@ -25,19 +25,19 @@ const useStyles = makeStyles(theme => ({
 
 const Prayers = props => {
   useRenderCounts("Prayers.js");
+
   const { prdata: data } = props;
   const [expanded, setExpanded] = useState(false);
   // Whatsapp logger
   const [log, setLogs] = useWhatsapplogger({});
   useEffect(() => {
-    if (expanded) {
-      // Whatsapp Logger
-      setLogs({
-        action: "Hijri Info",
-        message: `just opened up accordion`
-      });
-    }
-  }, [expanded]);
+    // Whatsapp Logger
+    setLogs({
+      action: "Prayer Times",
+      message: `just had a look at the prayer times of ${localStorage.getItem('padachone:place')} - ${localStorage.getItem('padachone:region')}`
+    });
+  }, [])
+  useHijriHook({expanded, travelorMain : 'Main'});
 
   const {
     data: { timings }