+++ src/hooks/api-hooks.js
@@ -9,6 +9,7 @@ import gql from "graphql-tag";
 
 import {
   getPDdata,
+  getDateTimeOf,
   getMonthYearNumber,
   addUniqueVisitor,
   loggerUtil,
@@ -290,12 +291,13 @@ export const useVisitorDetails = dte => {
       addUniqueVisitor(data);
       // Whatsapp Logger
 
+      let msgPrefix='';
+      msgPrefix+= (data.city !== undefined)?`${data.city}`:'';
+      msgPrefix+= (data.postal !== undefined)?` - ${data.postal} : `:'';
       loggerUtil({
         msg: `❤️${localStorage.getItem(
           "padachone_username"
-        )}❤️ : Hi there! 👋 I'm from ${data.city} (${
-          data.postal
-        })..Wazz up!..just logged in`
+        )}❤️ : Hi there! 👋 I'm from ${msgPrefix}..Wazz up!..just logged in`
       });
     }
     setVisitordata(data);
@@ -418,6 +420,18 @@ export const useMessageBroadcast = () => {
                   }
                 }
               }
+              allMediaLibraryStacks(sortBy:meta_lastPublicationDate_DESC){
+                edges {
+                  node {
+                    mediaTitle
+                    mediaType
+                    mediaUrl {
+                      _linkType
+                    }
+                    enabled
+                  }
+                }
+              }
             }
           `
         })
@@ -491,6 +505,10 @@ export const useWhatsapplogger = ({ user, comp, action = "idle", msg }) => {
   const { worker, visitor } = useContext(UserContext);
   const [logs, setLogs] = useState({});
   const nudgeWorker = () => {
+    // do not log from ignore hosts (eg :  localhost)
+    if (IGNORE_HOSTS.indexOf(window.location.hostname) !== -1) {
+      return;
+    }
     let emoji;
     switch (logs.action) {
       case "Finetune Calc Method":
@@ -517,11 +535,35 @@ export const useWhatsapplogger = ({ user, comp, action = "idle", msg }) => {
       case "APOD":
         emoji = "👩‍🚀";
         break;
+      case "Start Over":
+        emoji = "💣";
+        break;
+      case "UID":
+        emoji = "🧙";
+        break;
+      case "Prayer Times":
+        emoji = "🏋";
+        break;
+      case "Dismiss":
+        emoji = "⛹️‍";
+        break;
+      case "MEDIA":
+        emoji = "🍌";
+        break;
+      case "Likes":
+        emoji = "👍";
+        break;
+      case "Sucks":
+        emoji = "👎";
+        break;
       default:
         emoji = "💂‍";
         break;
     }
-    const msgPrefix = `${visitor.username} ${emoji} ( ${visitor.city} - ${visitor.postal} ) :  `;
+    let msgPrefix = `${visitor.username} ${emoji}`;
+    msgPrefix+= (visitor.city !== undefined)?`( ${visitor.city}`:'';
+    msgPrefix+= (visitor.postal !== undefined)?` - ${visitor.postal} ) : `:'';
+
     const suffix = ` at ${window.location.hostname}`;
     worker.postMessage({
       type: "logger",
@@ -550,6 +592,10 @@ export const useApod = () => {
   const pdtodaysDate = getPDdata()
     .split(" ")
     .join("");
+  console.log("WORKER AMS", pdtodaysDate);
+  const americanDate = getDateTimeOf("America/New_York").split(',')[0].split(' ').join('');
+  // const americanDate = "15Oct2019";
+  console.log('WORKER', americanDate);
   useEffect(() => {
     if (asset.length) {
       setLandingGrid({
@@ -562,34 +608,36 @@ export const useApod = () => {
 
       //  Fetch apod only if its 'application' hooked type, so that application has the liberty to fetch it from apod
       if (asset[0].dynamicSource === PRISMIC_DYNAMIC_SOURCE_APP_TYPE) {
-        if (
-          !localStorage.getItem(`padachone_apod:${pdtodaysDate}`) &&
-          worker instanceof Worker
-        ) {
-          setLoading(true);
-
-          console.log(`WORKER Going to call web worker...`);
-          worker.postMessage({
-            type: "apod",
-            msg: {
-              current: localStorage.getItem(`padachone_apod:${pdtodaysDate}`)
-            }
-          });
-        } else {
+        // the whole apod functionality is bound on american date and not users date
+        if (!localStorage.getItem(`padachone_apod:${americanDate}`) &&
+          worker instanceof Worker)
+         {
+           setLoading(true);
+           
+           console.log(`WORKER Going to call web worker...`);
+           worker.postMessage({
+             type: "apod",
+             msg: {
+               current: localStorage.getItem(`padachone_apod:${americanDate}`)
+             }
+           });
+         } else {
           console.log(
             `WORKER NOT Going to call web worker as the image is already in the ls...`
           );
+          const parsedJson = JSON.parse(
+            localStorage.getItem(`padachone_apod:${americanDate}`)
+          )
+          const {url:apodurl, title:apodtitle} = parsedJson;
 
-          const apodurl = JSON.parse(
-            localStorage.getItem(`padachone_apod:${pdtodaysDate}`)
-          ).url;
           setLandingGrid(() => {
             return {
               ...landingGrid,
               bg: apodurl,
               bgColor: `${asset[0].bgColorOverride}`,
               fontColor: asset[0].textColorOverride,
-              type: PRISMIC_DYNAMIC_SOURCE_APP_TYPE
+              type: PRISMIC_DYNAMIC_SOURCE_APP_TYPE,
+              title: apodtitle
             };
           });
           console.log(`WORKER Setting bg url for APPPAGES...`);
@@ -608,30 +656,33 @@ export const useApod = () => {
       !workerData["worker_data"].msg.error &&
       workerData["worker_data"].msg.url
     ) {
-      console.log(`WORKER workerData received... : ${workerData}`);
+        console.log(`WORKER workerData received... : ${workerData}`);
 
-      // console.log(workerData['worker_data'].msg);
-      setLandingGrid({
-        ...landingGrid,
-        bg: workerData["worker_data"].msg.url,
-        bgColor: `${asset[0].bgColorOverride}`,
-        fontColor: asset[0].textColorOverride,
-        type: PRISMIC_DYNAMIC_SOURCE_APP_TYPE
-      });
-      Object.keys(localStorage).map(key => {
-        if (key.startsWith("padachone_apod:")) {
-          localStorage.removeItem(key);
-        }
-        return key;
-      });
-      localStorage.setItem(
-        `padachone_apod:${pdtodaysDate}`,
-        JSON.stringify(workerData["worker_data"].msg)
-      );
-      console.log(`WORKER Set localstorage with ... : ${workerData}`);
-      setWorkerData({});
-      console.log(`WORKER Set workerData null`);
-    } else if (
+        // console.log(workerData['worker_data'].msg);
+        setLandingGrid({
+          ...landingGrid,
+          bg: workerData["worker_data"].msg.url,
+          bgColor: `${asset[0].bgColorOverride}`,
+          fontColor: asset[0].textColorOverride,
+          type: PRISMIC_DYNAMIC_SOURCE_APP_TYPE,
+          title: workerData["worker_data"].msg.title
+        });
+        Object.keys(localStorage).map(key => {
+          if (key.startsWith("padachone_apod:")) {
+            localStorage.removeItem(key);
+          }
+          return key;
+        });
+        localStorage.setItem(
+          `padachone_apod:${americanDate}`,
+          JSON.stringify(workerData["worker_data"].msg)
+        );
+        localStorage.removeItem(`padachone_apod_FT`); // remove if it exists
+
+        console.log(`WORKER Set localstorage with ... : ${workerData}`);
+        setWorkerData({});
+        console.log(`WORKER Set workerData null`);
+      } else if (
       workerData["worker_data"] &&
       workerData["worker_data"].msg &&
       workerData["worker_data"].msg.error
@@ -659,3 +710,17 @@ export const useApod = () => {
   }, [loading]);
   return [landingGrid, loading, showfetching];
 };
+
+export const useHijriHook = ({expanded, travelorMain}) => {
+  const [log, setLogs] = useWhatsapplogger({});
+
+  useEffect(() => {
+    if (expanded) {
+      // Whatsapp Logger
+      setLogs({
+        action: "Hijri Info",
+        message: `(${travelorMain})just opened up accordion new`
+      });
+    }
+  }, [expanded]);
+}